requirements.txt
openai-whisper== 20240930
flask==2.0.1
librosa==0.9.1
matplotlib==3.5.1
nltk==3.6.5
numpy>=1.20.0
soundfile>=0.10.3